Account overview page and a new templatetag
The account overview page is the start of the account renewal work.
We would like to have a page that allows a user to easily submit payments to renew/upgrade their account.

The templatetag allows for getattr functionality, this is useful when dealing with Managers and the first tag
(e.g. objects_set.all|first|getattr:'fieldname')